requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
aiohttp==3.9.5
flask==3.0.3
pillow==10.3.0
numpy==1.26.4